Fixed up LongTextWidget dependency issues
